requirements.txt
fbchat
pillow
python-dotenv